Moved Hermes packets body into specific folder (Help for import from GS)
