+++ data_logger/src/Sensors/HoneywellRscWrapper.hpp
@@ -19,7 +19,7 @@
 #include "globalVariables.hpp"
 #include "macrofunctions.h"
 
-#include "HoneywellRscBody.h"
+#include "PacketBody/HoneywellRscBody.h"
 
 // *************** HoneywellRsc Packets *************** //
