+++ ground_station/ERT2020GS/ext/Hermes_Packets
@@ -1,0 +1,1 @@
+C:/Users/maelf/Desktop/ENSL/EPFL/Master/Rocketry/SupersonicTestRocket/data_logger/src/Sensors/PacketBody/
