--- ground_station/ERT2020GS/ext/Hermes_Sensors
@@ -1,1 +1,0 @@
-C:/Users/maelf/Desktop/ENSL/EPFL/Master/Rocketry/SupersonicTestRocket/data_logger/src/Sensors/
